# docs/api/openapi-draft.yaml
openapi: 3.0.3
info:
  title: CareLink API
  description: 'This is a spec-first (design-first) API contract: it is hand-written and

    reviewed before the corresponding backend endpoint is implemented. It is

    the source of truth for the HTTP contract — controllers must match it,

    not the other way around.


    Only the `Auth` group is backed by a running controller today. Every

    other group is a draft: agreed enough to shape the frontend and the

    schema, but not yet implemented, and open to change as each feature

    module is actually built. Draft operations carry `x-status: draft` in

    this file; implemented ones carry `x-status: implemented`.

    '
  version: 0.0.1
  license:
    name: UNLICENSED
servers:
- url: /api
  description: Same-origin, relative to whichever host serves the front end.
...
paths:
...
  /incidents:
    get:
      tags:
      - Incidents
      summary: List incidents
      x-status: implemented
      description: Manager's queue.
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Size'
      - name: status
        in: query
        schema:
          $ref: '#/components/schemas/IncidentStatus'
      - name: severity
        in: query
        schema:
          type: string
          enum:
          - LOW
          - MEDIUM
          - HIGH
      responses:
        '200':
          description: Page of incidents.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
                  page:
                    type: integer
                  size:
                    type: integer
                  totalElements:
                    type: integer
    post:
      tags:
      - Incidents
      summary: Report a care exception
      x-status: draft
      description: 'UC-CG04. Also created automatically by the missed-check-in scheduled scan (`source: SYSTEM_MISSED_CHECKIN`)
        — a visit nobody checked in for is converted into an exception rather than left to vanish.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreateRequest'
      responses:
        '201':
          description: Incident opened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /incidents/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    get:
      tags:
      - Incidents
      summary: Get an incident and its full timeline
      x-status: implemented
      description: >-
        Who reported it, who responded, every notification attempt, and when the bound family member acknowledged
        it — this is the query that replaces the three-day chat-log reconstruction.



        For FAMILY, authorise access using the authenticated family identity, an ACTIVE and unexpired binding to
        the actual incident elder, and the applicable access scope. Derive the current familyMemberId from the
        session; the caller cannot choose another family member. The existence, delivery success or read state of
        a notification is not an authorisation condition. An otherwise authorised family member may read the
        incident and record their own view or acknowledgement even if no IN_APP notification exists or
        notification creation/delivery failed. Urgent alerts cannot be disabled through optional subscription
        preferences; that rule never overrides binding/access checks. The response window starts with the first
        successful urgent IN_APP notification creation for this incident/recipient and uses the configured
        interval (default two hours); further notices, polling and email retries do not reset it. GET
        /incidents/{id} returns acknowledgeBy, which is null when no response window has been established. A
        missing or failed notification does not block reading, viewing or acknowledging the incident. Later
        notification creation must preserve any already-recorded acknowledgement and must not restart waiting
        reminders for that family member. This recipient window is independent of manager incident status,
        respondBy and escalation. A family response contains safe incident facts, handling status and only the
        current family member's acknowledgement. Exclude raw coordinates, playbook, internal logs,
        responder/reporter user IDs and other recipients' acknowledgements. A GET is an access audit event, not
        proof of an explicit view or acknowledgement. The family page records a successful display with POST
        /incidents/{id}/view. Authorised non-family roles retain the full original incident timeline.


        Family callers are identified from the authenticated session, never a caller-supplied familyMemberId. For
        family callers, care resources require an ACTIVE binding to their actual elder, with no revocation and no
        elapsed expiresAt, and the applicable access scope at request time. Revocation/expiry takes effect on
        subsequent reads and writes, including historical content. Reading a family care resource records the
        caller, resource and server access time in the internal audit trail. Read access alone does not imply
        explicit acknowledgement. FULL/READ_ONLY do not imply different clinical field sets: both may use the
        family read projections below; an authorised family member may record their own incident
        view/acknowledgement under either scope because those commands record that family member's response, not a
        change to the elder care record. Notification read commands separately require ownership of the
        notification; no notification record is required for these incident operations.
      responses:
        '200':
          description: >-
            Family response uses the explicitly listed family schema. Other roles retain the original response
            schema; the server chooses by authenticated role, never by a client projection flag.
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/FamilyIncidentDetail'
                - $ref: '#/components/schemas/IncidentDetail'
        '401':
          $ref: '#/components/responses/FamilyUnauthorized'
        '403':
          $ref: '#/components/responses/FamilyForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      x-contract-status: specified
      x-family-use-cases:
      - UC-FM05
      x-family-owner: Member 4 (FM01-FM05 behaviour); source data remains with its assigned owner.
      security:
      - sessionCookie: []
      operationId: getIncident
      x-family-response-schema:
        $ref: '#/components/schemas/FamilyIncidentDetail'
  /incidents/{id}/claim:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      tags:
      - Incidents
      summary: Take over the incident
      x-status: implemented
      description: UC-MG05. Manager becomes `responderUserId`; starts the response countdown (`respondBy`).
      responses:
        '200':
          description: Claimed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '409':
          $ref: '#/components/responses/Conflict'
  /incidents/{id}/escalate:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      tags:
      - Incidents
      summary: Escalate to the next responder
      x-status: implemented
      description: 'UC-SYS02. The chain was assembled when the incident was opened (see `/incidents/{id}/escalation-chain`);
        this moves the incident to its next level: the new responder is set, both the previous and the new responder are notified,
        the trigger (`TIMEOUT` from the scheduled scan or `MANUAL` from a manager) and the reason are written to the timeline,
        and the next level''s countdown starts. A level whose responder is off duty is skipped and the skip is logged. The
        previous responder''s entry is never removed: responsibility accumulates, it does not transfer. When the chain is
        exhausted the incident becomes `UNRESOLVED_ESCALATED`, is pinned to the top-level view and the family is notified.

        '
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Free text; the scheduled scan writes "response countdown expired".
      responses:
        '200':
          description: Escalated; new `responderUserId` and `respondBy` set, or `UNRESOLVED_ESCALATED` if the chain is exhausted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '409':
          description: The incident is no longer waiting for a responder (already claimed or resolved) — the "almost escalated"
            case, which is logged.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /incidents/{id}/escalation-chain:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    get:
      tags:
      - Incidents
      summary: The escalation chain assembled for this incident
      x-status: implemented
      description: 'UC-MG05 / UC-SYS02. Shows the chain exactly as it was assembled at runtime for this incident — each level''s
        responder, that level''s countdown, and which level is current. Two incidents of the same category can have different
        chains (night vs day, who is on duty), which is why the chain is read per incident rather than configured.

        '
      responses:
        '200':
          description: Ordered levels, first responder first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscalationChain'
        '404':
          $ref: '#/components/responses/NotFound'
  /incidents/{id}/contact-attempts:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      tags:
      - Incidents
      summary: Record an attempt to contact the family
      x-status: implemented
      description: 'UC-MG05 step 4. Every attempt is recorded whether or not the family was reached, with the channel and
        the reason when not. A `NOT_REACHED` outcome is what lets the manager apply the standard playbook without waiting
        (4a); when every channel fails the timeline shows "family not reached" and the dashboard asks for manual contact (4b).

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactAttemptRequest'
      responses:
        '201':
          description: Attempt appended to the timeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentLogEntry'
        '400':
          $ref: '#/components/responses/ValidationError'
  /incidents/{id}/playbook:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      tags:
      - Incidents
      summary: Apply a standard response playbook
      x-status: implemented
      description: 'UC-MG05 step 5 / 4a. When the family cannot be reached, or agrees to it, the manager confirms the standard
        playbook for the incident''s category and severity; the steps it prescribes are appended to the timeline as the manager
        carries them out. The playbook is chosen from `/incident-playbooks`.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - playbookCode
              properties:
                playbookCode:
                  type: string
                  example: FALL_MEDIUM
      responses:
        '200':
          description: Playbook recorded on the incident.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /incidents/{id}/severity:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      tags:
      - Incidents
      summary: Raise or lower the severity
      x-status: implemented
      description: 'UC-MG05 5a. When the situation worsens the manager raises the severity; the escalation chain is re-assembled
        for the new severity and the countdown restarted, but the same incident and the same timeline continue — no new incident
        is opened.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - severity
              properties:
                severity:
                  $ref: '#/components/schemas/IncidentSeverity'
                reason:
                  type: string
      responses:
        '200':
          description: Severity changed; chain re-assembled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /incident-playbooks:
    get:
      tags:
      - Incidents
      summary: Standard response playbooks
      x-status: implemented
      description: The organisation's standard responses per category and severity, as configured (not code constants).
      parameters:
      - name: category
        in: query
        schema:
          $ref: '#/components/schemas/IncidentCategory'
      - name: severity
        in: query
        schema:
          $ref: '#/components/schemas/IncidentSeverity'
      responses:
        '200':
          description: Matching playbooks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncidentPlaybook'
  /incidents/{id}/resolve:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    post:
      tags:
      - Incidents
      summary: Resolve the incident
      x-status: implemented
      description: 'UC-MG05 step 6. Only the current responder may resolve. An incident is never closed automatically: without
        this call it stays open and keeps escalating.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - resolutionNote
              properties:
                resolutionNote:
                  type: string
                outcome:
                  type: string
                  enum:
                  - HANDLED_ON_SITE
                  - REFERRED_TO_MEDICAL_CARE
                  - FALSE_ALARM
      responses:
        '200':
          description: Resolved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '403':
          description: The caller is not the current responder.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '409':
          description: Already resolved.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: 'Server-side session established by `POST /auth/login`. State-changing requests must also echo the `XSRF-TOKEN`
        cookie value in an `X-XSRF-TOKEN` header.

        '
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  schemas:
...
    ProblemDetail:
      description: RFC 9457 `application/problem+json` body. Every error response uses this shape.
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
      additionalProperties: true
...
    IncidentSource:
      type: string
      enum:
      - CAREGIVER
      - ELDER_SOS
      - SYSTEM_MISSED_CHECKIN
    IncidentCategory:
      type: string
      enum:
      - SOS
      - MEDICAL
      - FALL
      - SERVICE
      - OTHER
    IncidentSeverity:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
    IncidentStatus:
      type: string
      enum:
      - OPEN
      - ACKNOWLEDGED
      - IN_PROGRESS
      - RESOLVED
      - UNRESOLVED_ESCALATED
    Incident:
      type: object
      required:
      - id
      - elderId
      - source
      - severity
      - status
      properties:
        id:
          type: integer
          format: int64
        elderId:
          type: integer
          format: int64
        visitId:
          type: integer
          format: int64
          nullable: true
        reportedByUserId:
          type: integer
          format: int64
          nullable: true
        responderUserId:
          type: integer
          format: int64
          nullable: true
        source:
          $ref: '#/components/schemas/IncidentSource'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        latitude:
          type: number
          format: double
          nullable: true
        longitude:
          type: number
          format: double
          nullable: true
        locationText:
          type: string
          nullable: true
        description:
          type: string
        respondBy:
          type: string
          format: date-time
          nullable: true
          description: Countdown deadline; past this, the scheduled job escalates automatically.
        reportedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    IncidentCreateRequest:
      type: object
      required:
      - elderId
      - category
      - severity
      properties:
        elderId:
          type: integer
          format: int64
        visitId:
          type: integer
          format: int64
        category:
          $ref: '#/components/schemas/IncidentCategory'
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        description:
          type: string
    IncidentLogEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        actor:
          type: string
        action:
          type: string
        detail:
          type: string
        occurredAt:
          type: string
          format: date-time
    IncidentAcknowledgement:
      type: object
      properties:
        id:
          type: integer
          format: int64
        incidentId:
          type: integer
          format: int64
        familyMemberId:
          type: integer
          format: int64
        viewedAt:
          type: string
          format: date-time
          nullable: true
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
        responseNote:
          type: string
          nullable: true
    IncidentDetail:
      allOf:
      - $ref: '#/components/schemas/Incident'
      - type: object
        properties:
          playbookCode:
            type: string
            nullable: true
          log:
            type: array
            items:
              $ref: '#/components/schemas/IncidentLogEntry'
          acknowledgements:
            type: array
            items:
              $ref: '#/components/schemas/IncidentAcknowledgement'
          timeline:
            type: array
            description: Handling history for the incident.
            items:
              $ref: '#/components/schemas/IncidentLogEntry'
...
    EscalationChain:
      type: object
      description: 'The escalation chain as it would be walked right now.

        It is assembled per request from severity, the hour of the day and the duty roster,
        and is deliberately not stored: asking twice can legitimately give different answers
        if the roster changed in between, which is why `assembledAt` is part of the answer.
        What happened, as opposed to what would happen, is in the incident timeline.

        '
      properties:
        incidentId:
          type: integer
          format: int64
        assembledAt:
          type: string
          format: date-time
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        assembledFrom:
          type: string
          description: The inputs the chain was built from, in words, for the manager's screen.
          example: severity HIGH, duty manager on shift, 3 manager(s) enabled
        levels:
          type: array
          items:
            $ref: '#/components/schemas/EscalationLevel'
    ContactAttemptRequest:
      type: object
      required:
      - channel
      - outcome
      properties:
        channel:
          type: string
          enum:
          - PHONE
          - PUSH
          - EMAIL
        outcome:
          type: string
          enum:
          - REACHED
          - NOT_REACHED
        note:
          type: string
          description: What was agreed, or why the family could not be reached.
    IncidentPlaybook:
      type: object
      description: 'The standard response a manager falls back on when the family cannot be
        reached. One per incident category; the manager does not wait for the family before
        acting.

        '
      properties:
        code:
          type: string
          example: PB-SOS
        category:
          $ref: '#/components/schemas/IncidentCategory'
        title:
          type: string
        steps:
          type: array
          items:
            type: string
...
    FamilyProblemDetail:
      type: object
      required:
      - type
      - title
      - status
      description: >-
        Error contract for FM01-FM05. Optional code is a stable machine-readable reason; fields contains input
        validation messages.
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
      additionalProperties: true
      example:
        type: about:blank
        title: Insufficient permission
        status: 403
        detail: The current family member cannot access this resource.
...
    FamilyIncidentAcknowledgement:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          nullable: true
          description: >-
            Null in a detail read before a view/acknowledgement record is created. Non-null after a successful
            view or acknowledge command.
        incidentId:
          type: integer
          format: int64
          minimum: 1
        familyMemberId:
          type: integer
          format: int64
          minimum: 1
        viewedAt:
          type: string
          format: date-time
          nullable: true
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
        responseNote:
          type: string
          maxLength: 255
          nullable: true
      description: >-
        Current family member only. Viewed and acknowledged timestamps are independent and preserve their first
        server-recorded values.
      required:
      - id
      - incidentId
      - familyMemberId
      - viewedAt
      - acknowledgedAt
      - responseNote
    FamilyIncidentDetail:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
        elderId:
          type: integer
          format: int64
        visitId:
          type: integer
          format: int64
          nullable: true
        source:
          $ref: '#/components/schemas/IncidentSource'
        category:
          $ref: '#/components/schemas/IncidentCategory'
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        description:
          type: string
          nullable: true
        reportedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        acknowledgeBy:
          type: string
          format: date-time
          nullable: true
          description: >-
            This family member response deadline: first successfully created urgent IN_APP notification for this
            incident/family member plus the configured window (default two hours). Null when no response window
            was established, including when notification creation has not succeeded; null does not prevent an
            authorised read, view or acknowledgement. Stable across polling, email retries and further
            notifications for the same incident. Never the manager respondBy deadline. Later notification creation
            does not clear an existing acknowledgement or reopen waiting reminders.
        acknowledgement:
          $ref: '#/components/schemas/FamilyIncidentAcknowledgement'
      description: >-
        Safe incident facts and global handling status for the currently authorised family member. Notification
        existence or delivery is not an access prerequisite. No raw location coordinates, playbook, internal log,
        staff account IDs or other family acknowledgements.
      required:
      - id
      - elderId
      - visitId
      - source
      - category
      - severity
      - status
      - description
      - reportedAt
      - resolvedAt
      - acknowledgeBy
      - acknowledgement
...
  responses:
...
    NotFound:
      description: No resource with this id.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: The request is well-formed but illegal in the resource's current state (e.g. an out-of-order visit transition).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    ValidationError:
      description: Request failed validation.
      content:
        application/problem+json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ProblemDetail'
            - type: object
              properties:
                fields:
                  type: object
                  additionalProperties:
                    type: string
                  description: Field name to validation message.
    FamilyUnauthorized:
      description: No valid authenticated session.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FamilyProblemDetail'
    FamilyForbidden:
      description: >-
        The current caller lacks the family role, resource ownership, recipient identity, or a currently active,
        unexpired binding/access scope. An authenticated caller does not gain access by changing an ID.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/FamilyProblemDetail'
...
security:
- sessionCookie: []